Streamline Docker setup instructions
Most of the page was selling Docker as a technology instead of telling you how to get set up.

On top of that it made questionable claims. E.g. praising how it didn't use an expensive VM, even though the majority of users would have an Alpine Linux VM acting as a container host ( Hyperkit hypervisor on Mac, Hyper-V on Windows ).

This PR removes the text ass it adds very little to the document.